Reorganized CalculateSuitLength module
